refactor: 패키지명을 api -> in, spi -> out으로 변경한다
