+++ app/src/test/java/shop/woowasap/accept/product/ProductFixture.java
@@ -2,8 +2,8 @@ package shop.woowasap.accept.product;
 
 import java.time.LocalDateTime;
 import shop.woowasap.auth.controller.request.LoginRequest;
-import shop.woowasap.shop.domain.api.product.request.RegisterProductRequest;
-import shop.woowasap.shop.domain.api.product.request.UpdateProductRequest;
+import shop.woowasap.shop.domain.in.product.request.RegisterProductRequest;
+import shop.woowasap.shop.domain.in.product.request.UpdateProductRequest;
 
 public class ProductFixture {
