+++ coverage-exclude.asap
@@ -1,26 +1,9 @@
 exclude shop/woowasap/Application
-exclude shop/woowasap/auth/domain/*
-exclude shop/woowasap/*/domain/api/*/request/*
-exclude shop/woowasap/*/domain/api/*/response/*
-exclude shop/woowasap/*/domain/spi/request/*
-exclude shop/woowasap/*/domain/spi/response/*
-exclude shop/woowasap/*/domain/exception/*
-exclude shop/woowasap/*/controller/*
-exclude shop/woowasap/*/repository/*
-exclude shop/woowasap/*/repository/entity/*
-exclude shop/woowasap/*/app/api/request/*
-exclude shop/woowasap/*/app/api/response/*
-exclude shop/woowasap/*/app/spi/request/*
-exclude shop/woowasap/*/app/spi/response/*
-exclude shop/woowasap/*/app/exception/*
-exclude shop/woowasap/order/domain/exception/*
-exclude shop/woowasap/order/domain/in/request/*
-exclude shop/woowasap/order/controller/*
-exclude shop/woowasap/order/controller/request/*
-exclude shop/woowasap/order/repository/*
-exclude shop/woowasap/order/payment/random/RandomOrderPayment
 exclude shop/woowasap/*/domain/*/request/*
 exclude shop/woowasap/*/domain/*/response/*
+exclude shop/woowasap/*/domain/exception/*
+exclude shop/woowasap/*/controller/*
+exclude shop/woowasap/*/controller/request/*
 exclude shop/woowasap/auth/service/configuration/*
 exclude shop/woowasap/auth/infra/*/*
-exclude shop/woowasap/*/controller/request/*
+exclude shop/woowasap/order/payment/random/RandomOrderPayment
