+++ order/order-service/src/test/java/shop/woowasap/order/service/OrderServiceTest.java
@@ -4,14 +4,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchException;
 import static org.mockito.Mockito.when;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
@@ -35,8 +33,8 @@ import shop.woowasap.order.service.support.fixture.OrderDtoFixture;
 import shop.woowasap.order.service.support.fixture.OrderFixture;
 import shop.woowasap.order.service.support.fixture.OrderProductFixture;
 import shop.woowasap.order.service.support.fixture.ProductFixture;
-import shop.woowasap.shop.domain.api.cart.CartConnector;
-import shop.woowasap.shop.domain.api.product.ProductConnector;
+import shop.woowasap.shop.domain.in.cart.CartConnector;
+import shop.woowasap.shop.domain.in.product.ProductConnector;
 import shop.woowasap.shop.domain.cart.Cart;
 import shop.woowasap.shop.domain.cart.CartProduct;
 import shop.woowasap.shop.domain.product.Product;
