+++ shop/shop-domain/src/main/java/shop/woowasap/shop/domain/in/cart/response/CartProductResponse.java
@@ -1,4 +1,4 @@
-package shop.woowasap.shop.domain.api.cart.response;
+package shop.woowasap.shop.domain.in.cart.response;
 
 public record CartProductResponse(long productId, String name, String price, long quantity) {
