+++ shop/shop-service/src/main/java/shop/woowasap/shop/service/ProductConnectorService.java
@@ -4,9 +4,9 @@ import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shop.woowasap.shop.domain.api.product.ProductConnector;
+import shop.woowasap.shop.domain.in.product.ProductConnector;
 import shop.woowasap.shop.domain.product.Product;
-import shop.woowasap.shop.domain.spi.ProductRepository;
+import shop.woowasap.shop.domain.out.ProductRepository;
 
 @Service
 @RequiredArgsConstructor
